third_party/blink/web_tests/storage/domstorage/sessionstorage/string-conversion-expected.txt

This test case verifies that session storage only stores strings.
Length is 0
Testing implicit setters
Type/value for null is string/null
Type/value for 0 is string/0
Type/value for function(){} is string/function(){}
Testing explicit setters
Type/value for null is string/null
Type/value for 0 is string/0
Type/value for function(){} is string/function(){}
Testing index setters
Type/value for null is string/null
Type/value for 0 is string/0
Type/value for function(){} is string/function(){}
